rooftop/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-filter==2.3.0
django-rest-swagger==2.1.0
djangorestframework==3.11.1
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2-binary==2.8.5
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
zipp==3.1.0
djangorestframework-gis==0.14
djangorestframework-simplejwt==4.3.0
django-cors-headers==3.4.0
dj-database-url
gunicorn
Pillow==7.2.0
pandas==1.1.2
django-import-export==2.3.0